refactor: move channels_v2/channel_base.py to channels/channel_base.py
