+++ apps/channels/channels_v2/api_channel.py
@@ -4,7 +4,7 @@ from typing import TYPE_CHECKING
 
 from apps.channels.callbacks import ChannelCallbacks
 from apps.channels.capabilities import ChannelCapabilities
-from apps.channels.channels_v2.channel_base import ChannelBase
+from apps.channels.channel_base import ChannelBase
 from apps.channels.const import MESSAGE_TYPES
 from apps.channels.pipeline import MessageProcessingContext, MessageProcessingPipeline
 from apps.channels.sender import ChannelSender
