+++ apps/channels/channels_v2/sureadhere_channel.py
@@ -5,7 +5,7 @@ from typing import TYPE_CHECKING
 
 from apps.channels.callbacks import ChannelCallbacks
 from apps.channels.capabilities import ChannelCapabilities
-from apps.channels.channels_v2.channel_base import ChannelBase
+from apps.channels.channel_base import ChannelBase
 from apps.channels.models import ChannelPlatform
 from apps.channels.sender import ChannelSender
