+++ apps/channels/channels_v2/whatsapp_channel.py
@@ -6,7 +6,7 @@ from typing import TYPE_CHECKING
 
 from apps.channels.callbacks import ChannelCallbacks
 from apps.channels.capabilities import ChannelCapabilities
-from apps.channels.channels_v2.channel_base import ChannelBase
+from apps.channels.channel_base import ChannelBase
 from apps.channels.datamodels import WhatsAppMessage
 from apps.channels.models import ChannelPlatform
 from apps.channels.sender import ChannelSender
